+++ include/nuttx/fs/fs.h
@@ -597,7 +597,7 @@ int register_blockdriver(FAR const char *path,
 #ifndef CONFIG_DISABLE_MOUNTPOINT
 int register_blockpartition(FAR const char *partition,
                             mode_t mode, FAR const char *parent,
-                            size_t firstsector, size_t nsectors);
+                            off_t firstsector, off_t nsectors);
 #endif
 
 /****************************************************************************
@@ -674,7 +674,7 @@ int register_mtddriver(FAR const char *path, FAR struct mtd_dev_s *mtd,
 #ifdef CONFIG_MTD
 int register_mtdpartition(FAR const char *partition,
                           mode_t mode, FAR const char *parent,
-                          off_t firstblock, size_t nblocks);
+                          off_t firstblock, off_t nblocks);
 #endif
 
 /****************************************************************************
